backend/pyproject.toml
[tool.poetry]
name = "rented_rooms"
version = "0.1.0"
description = ""
authors = ["teddy <teddylee777@gmail.com>"]
license = "Apache 2.0 License"
readme = "README.md"

packages = [
    { include = "graphparser" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
langchain = "^0.3"
langchain-openai = "^0.2"
langsmith = "^0.1.115"
langgraph = "^0.2.19"
pymupdf = "^1.24.10"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
ipython = "^8.27.0"
faiss-cpu = "^1.8.0.post1"
beautifulsoup4 = "^4.12.3"
markdownify = "^0.13.1"
pdfplumber = "^0.11.4"
lxml = "^5.3.0"
pypdf = "^4.2.0"
pandas = "^2.2.2"
rank-bm25 = "^0.2.2"
redis = "^5.0.3"
chromadb = "^0.4.24"
pdfminer-six = "^20231228"
pymupdf4llm = "^0.0"
matplotlib = "^3.8.4"
streamlit = "^1.32.2"
jupyter = "^1.0.0"
notebook = "^7.1.2"
torch = "^2.2.2"
torchvision = "^0.17.2"
open-clip-torch = "^2.24.0"
pydantic = "^2.7.4"
lark = "^1.1.9"
unstructured = { version = "0.10.19", extras = ["all-docs"] }
arxiv = "^2.1.0"
tiktoken = "^0.7"
tenacity = "8.3.0"
pymilvus = "^2.4.4"
google-search-results = "^2.4.2"
protobuf = "3.20.3"
sqlalchemy = "^2.0.31"
llama-index-core = "^0.10.58"
llama-parse = "^0.4.9"
llama-index-readers-file = "^0.1.30"
flashrank = "^0.2.8"
docx2txt = "^0.8"
nest-asyncio = "^1.6.0"
rapidocr-onnxruntime = "^1.3.24"
seaborn = "^0.13.2"
grandalf = "^0.8"
rouge-score = "^0.1.2"
mypy = "^1.11.2"
pinecone = "^5.3.1"
wikipedia = "^1.4.0"
scikit-learn = "^1.5.2"
huggingface-hub = "^0.25"
openai = "^1.34"
deepl = "^1.17.0"
kiwipiepy = "^0.17.1"
konlpy = "^0.6.0"
langchain-chroma = "^0.1.4"
langchain-huggingface = "^0.1.2"
ragas = "^0.2.9"
langchain-google-genai = "^2.0.7"
boto3 = "^1.35.91"
geopy = "^2.4.1"
pypdf2 = "^3.0.1"
langchain-teddynote = "^0.3.41"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
